compile fixes, warnings, including the new int vs uint for count() etc
git-svn-id: svn://scribus.net/branches/Version135x/Scribus@8545 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/colorchart.cpp
@@ -36,7 +36,7 @@ for which a new license (GPL+exception) is in place.
 ColorChart::ColorChart(QWidget *parent, ScribusDoc* doc) : QLabel(parent), m_doc(doc)
 {
 	setScaledContents( true );
-	setAlignment( static_cast<int>( Qt::AlignCenter ) );
+	setAlignment(Qt::AlignCenter);
 	Xp = 0;
 	Yp = 0;
 	setBackgroundMode(Qt::NoBackground);
